new math

+++ 0x00-math_complex/holberton.h
@@ -4,8 +4,8 @@
 #include <stdio.h>
 
 typedef struct complex{
-	int re;
-	int im;
+	double re;
+	double im;
 } complex;
 
 void display_complex_number(complex c);
